requirements.txt
Django==1.11
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.1.0
inflection==0.3.1
pytz==2018.4
